pyproject.toml
[project]
name = "tts-piper"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "HipsterBrown", email = "headhipster@hipsterbrown.com" }
]
dependencies = [
    "viam-sdk>=0.15.0",
    "speech-service-api @ git+https://github.com/viam-labs/speech-service-api.git@v0.4.0",
    "piper-tts>=1.2.0; platform_system == 'Linux'",
    "pyaudio>=0.2.14",
]
readme = "README.md"
requires-python = ">= 3.8"
license = { text = "Apache-2.0" }
classifiers = ["Private :: Do Not Upload"]

[project.scripts]
"tts-piper" = "tts_piper:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tts_piper"]
